+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/sync/SyncConfiguration.kt
@@ -551,7 +551,8 @@ public interface SyncConfiguration : Configuration {
                 compactOnLaunchCallback,
                 null, // migration is not relevant for sync,
                 initialDataCallback,
-                partitionValue == null
+                partitionValue == null,
+                durability
             )
 
             return SyncConfigurationImpl(